Commit message:
Add following users, tweeting, and unfollowing users feature

This commit adds the ability for users to follow and unfollow other users, and also post tweets in the Twitter-Clone application.
The functionality for following and unfollowing users is handled in the UserFollow model and UserRepository.
The function for posting tweets is handled in the Tweet model and TweetRepository.
Also, UserController and TweetController have endpoints for these features.
Next step is to lock the other endpoints behind authentication but for now its time for bed!