Include related user data when fetching tweets
The existing implementation of the tweet retrieval methods in the TweetRepository were modified to include related user data. This was done to ensure that consumers of the API, particularly the client-side application, do not need to make additional requests to get user details when displaying tweets. The change affects the methods for getting a single tweet by ID, all tweets, and tweets by a specific user's ID.