Add profile image path to User model
A nullable `ProfileImagePath` property was added to the `User` model to allow storing the file path for user profile photos. The corresponding migration was created. An `ImagesController` was also created to handle the image uploads. Users can now upload profile pictures and their paths will be stored in the database for future retrieval. The `CurrentuserService.ts` was updated to reflect these changes. I have used Six Labors ImageSharp library to accept an incoming image from the front end, create a local directory based off of the current users username and convert the uploaded file from the user to a optimized webp for later retrieval and display on the front end.