This commit includes the addition of TweetDisplay and HomeComponent modules for displaying tweets and home components respectively.
Tweets from people you follow now display on the homepage which you are brought to upon login. Created a custom pipe to handle display of "Time Since Tweet Created" on Tweet itself. Added fallback images for a default profile icon svg if a user has not uploaded a profile photo yet. Also modified TweetDto to contain a username field to make requesting a profile photo on the tweets themselves simpler.

- Currently designing the tweet display itself.